Fix Back-End: Tag Search
Fix S03P13A509-112
Tag검색 되도록 오류 수정
avgScore DB에서 바로 받아 이용
avgScore float형으로 변경
정렬 출력 기능 Java에서 계산 후 반환

추후 nickname 검색 확인 필요, video type, style등 복합 검색 테스트 필요
+ auto failover 필요

+++ backend/src/main/java/com/web/editor/model/dto/user/search/SearchPortfolioJoinBookmark.java
@@ -23,7 +23,7 @@ public class SearchPortfolioJoinBookmark {
     private String url;
     // private String skill;
     private Integer bookmarkNumber;
-    private Integer avgScore;
+    private float avgScore;
     private String tagKey;
 
     public void setInit(String url, String tagKey){
